Version 2.1.3
Added support for option --ignore-unions. Anonymous unions cannot be
transformed to named unions in case the union's data members are used inside
member functions of a class but outside the named union scope.
This is a temporary option that allows to ignore all unions as a workaround
when this transformation brakes codes.